refactor: use streaming for refs local
