+++ src/core/components/files-regular/refs-local-pull-stream.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const CID = require('cids')
+const base32 = require('base32.js')
 const pull = require('pull-stream')
 const pullDefer = require('pull-defer')
 
@@ -7,10 +9,26 @@ module.exports = function (self) {
   return () => {
     const deferred = pullDefer.source()
 
-    self.refs.local()
-      .catch((err) => deferred.resolve(pull.error(err)))
-      .then((refs) => deferred.resolve(pull.values(refs)))
+    self._repo.blocks.query({ keysOnly: true }, (err, blocks) => {
+      if (err) {
+        return deferred.resolve(pull.error(err))
+      }
+
+      const refs = blocks.map(b => dsKeyToRef(b.key))
+      deferred.resolve(pull.values(refs))
+    })
 
     return deferred
+  }
+}
+
+function dsKeyToRef (key) {
+  // Block key is of the form /<base32 encoded string>
+  const decoder = new base32.Decoder()
+  const buff = decoder.write(key.toString().slice(1)).finalize()
+  try {
+    return { ref: new CID(buff).toString() }
+  } catch (err) {
+    return { err: `Could not convert block with key '${key}' to CID: ${err.message}` }
   }
 }
